requirements.txt
hodgeys_kjv_db
tweepy
pillow